Refactor: Enhance Pile Creation and Editing with Icon and Color Selection
This commit significantly refactors the pile creation and editing functionality by introducing icon and color selection.

**Key Changes:**

- **`AlertNewPile.kt` & `AlertEditPile.kt`:**
    - Unified into a private `AlertNewEditPile` composable.
    - Users can now select an icon and a color for new or existing piles.
    - `AlertNewPile` now calls `onConfirm` with `pileName`, `pileIconId`, and `pileColorNumber`.
    - `AlertEditPile` is a new composable that allows editing pile name, icon, and color. It's now used in `PileDetailScreen`.
    - The alert dialog UI now includes:
        - A preview of the pile icon with its selected color.
        - A `ToggleButton` group to switch between "Color" and "Icon" selection.
        - A `LazyVerticalGrid` to display available colors or icons based on the `ToggleButton` selection.
        - Selected color/icon is highlighted.
- **ViewModels (`AddDocumentViewModel`, `EditDocumentPilesViewModel`, `HomeViewModel`, `PileDetailViewModel`):**
    - `addPile` methods updated to accept `iconId` and `color` parameters.
    - `PileDetailViewModel.updatePileName` renamed to `updatePileName` and now also accepts `newPileIconId` and `newPileColor`.
- **Screens (`AddDocumentScreen`, `EditDocumentPilesScreen`, `HomeScreen`, `PileDetailScreen`):**
    - Updated to pass `pileIconId` and `pileColorNumber` when calling `addPile` or `updatePileName` in their respective ViewModels.
    - `PileDetailScreen` now uses the new `AlertEditPile` composable.
- **`PileModelRepository.kt`:**
    - `insertPileModel` and `updatePileModel` now trim whitespace from the `pileModel.name` before database operations.
- **`strings.xml`:**
    - Added "Color" and "Icon" string resources.
- **Project Files:**
    - Added `.idea/dictionaries/project.xml` to the project.

This refactoring provides a more visually rich and customizable experience for managing piles.

+++ app/src/main/java/com/ganadoro/pile/ui/screens/home/HomeViewModel.kt
@@ -72,13 +72,13 @@ class HomeViewModel(
         bitmapCacheRepository.ensureBitmapIsLoaded(documentId)
     }
 
-    fun addPile(pileName: String, iconId: String, color: Int) {
+    fun addPile(pileName: String, iconId: String, color: Long) {
         viewModelScope.launch {
             val newPile = PileModel(
                 id = UUID.randomUUID().toString(),
                 name = pileName,
                 iconId = iconId,
-                colorNumber = color.toLong()
+                colorNumber = color
             )
 
             pileModelRepository.insertPileModel(newPile)
